requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.31.0
plotly>=5.18.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
prometheus_client>=0.20.0
kubernetes>=28.1.0
pyyaml>=6.0
